+++ src/screens/auth/password/steps/VerificationStep.tsx
@@ -1,9 +1,8 @@
 import styled from '@emotion/native';
-import DefaultInput from "@/components/auth/authInput/DefaultInput"
+import DefaultInput from "@/components/common/input/DefaultInput"
 import Typography from "@/components/common/typography/Typography"
 import { useEffect, useState } from "react";
 import ContentLayout from '@/components/layout/authLayout/AuthStepComponentLayout/ContentLayout';
-import { Text } from 'react-native';
 import { isValidEmail } from '@/utils/isValidEmail';
 
 interface VerificationStepProps {
