+++ src/screens/auth/signup/steps/EmailStep.tsx
@@ -1,4 +1,4 @@
-import DefaultInput from "@/components/auth/authInput/DefaultInput";
+import DefaultInput from "@/components/common/input/DefaultInput";
 import Typography from "@/components/common/typography/Typography";
 import { useEffect, useState } from "react";
 import ContentLayout from "@/components/layout/authLayout/AuthStepComponentLayout/ContentLayout";
