+++ src/screens/auth/signup/steps/VerificationStep.tsx
@@ -1,5 +1,5 @@
 import styled from '@emotion/native';
-import DefaultInput from "@/components/auth/authInput/DefaultInput"
+import DefaultInput from "@/components/common/input/DefaultInput"
 import Typography from "@/components/common/typography/Typography"
 import { useEffect } from "react";
 import ContentLayout from '@/components/layout/authLayout/AuthStepComponentLayout/ContentLayout';
